+++ src/org/previmer/ichthyop/dataset/Mars2dOpendapDataset.java
@@ -26,7 +26,7 @@ public class Mars2dOpendapDataset extends Mars2dCommon {
 
     @Override
     void setOnFirstTime() throws Exception {
-        long t0 = getSimulationManager().getTimeManager().get_tO();
+        double t0 = getSimulationManager().getTimeManager().get_tO();
         DatasetIO.checkInitTime(ncIn, strTime);
         rank = findCurrentRank(t0);
         time_tp1 = t0;
@@ -34,7 +34,7 @@ public class Mars2dOpendapDataset extends Mars2dCommon {
 
     @Override
     public void nextStepTriggered(NextStepEvent e) throws Exception {
-        long time = e.getSource().getTime();
+        double time = e.getSource().getTime();
         //Logger.getAnonymousLogger().info("set fields at time " + time);
         int time_arrow = (int) Math.signum(e.getSource().get_dt());
