+++ src/org/previmer/ichthyop/dataset/Mars3dOpendapDataset.java
@@ -20,7 +20,7 @@ public class Mars3dOpendapDataset extends Mars3dCommon {
 
     @Override
     void setOnFirstTime() throws Exception {
-        long t0 = getSimulationManager().getTimeManager().get_tO();
+        double t0 = getSimulationManager().getTimeManager().get_tO();
         DatasetIO.checkInitTime(ncIn, strTime);
         rank = findCurrentRank(t0);
         time_tp1 = t0;
@@ -28,7 +28,7 @@ public class Mars3dOpendapDataset extends Mars3dCommon {
 
     @Override
     public void nextStepTriggered(NextStepEvent e) throws Exception {
-        long time = e.getSource().getTime();
+        double time = e.getSource().getTime();
         //Logger.getAnonymousLogger().info("set fields at time " + time);
         int time_arrow = (int) Math.signum(e.getSource().get_dt());
